requirements.txt
asttokens==2.4.1
colorama==0.4.6
executing==2.1.0
icecream==2.1.3
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
Pygments==2.18.0
pyqtdarktheme==0.1.7
PySide6_Essentials==6.8.0.1
pytest==8.3.4
shiboken6==6.8.0.1
six==1.16.0
style==1.1.0
update==0.0.1
